FIX: Refactor imports in HeaderCell.tsx and PasteForm.tsx to use TableApiType from common types
